scraping/requirements.txt
requests
psycopg2-binary
tqdm